added/stole and modified a frontend tutorial
I just wanted to try and see if i could follow a tutorial and add a front end . so far its only functionality is login and registration. I have no idea how to connect it to our backend, this tutorial had a fake backend to test login and register functionality. Also might not want to use cause its basically copied with some changes. i left the hyper link to the tutorial so you can look if u want to. To run cd front end, npm install (dunno if this is necassary everytime) and npm runcdev to open it on a local host.